# docs/swagger.yaml
basePath: /
definitions:
  models.LoginResponse:
    properties:
      access_token:
        type: string
    type: object
  models.Product:
    properties:
      id:
        type: integer
      name:
        minLength: 3
        type: string
      price:
        type: integer
      stock:
        minimum: 0
        type: integer
    required:
    - name
    - price
    - stock
    type: object
  models.User:
    properties:
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      role:
        type: string
    type: object
  utils.APIResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: ahmad@example.com
    name: Ahmadchoms
  description: API untuk manajemen inventory warung (Backend Engineering v2).
  termsOfService: http://swagger.io/terms/
  title: Inventory API
  version: "2.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Mengecek email & password, lalu mengembalikan token JWT
      parameters:
      - description: Email dan Password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Masuk ke dalam sistem
      tags:
      - Auth
  /products:
    get:
      consumes:
      - application/json
      description: Mengambil list produk dengan pagination & search
      parameters:
      - description: Halaman ke- (Default 1)
        in: query
        name: page
        type: integer
      - description: Jumlah data (Default 10)
        in: query
        name: limit
        type: integer
      - description: Cari nama produk
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Ambil Semua Produk
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Menambahkan data produk ke database
      parameters:
      - description: Data Produk
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden (Bukan Admin)
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Tambah Produk Baru (Admin Only)
      tags:
      - Products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus produk dari database
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Hapus Produk (Admin Only)
      tags:
      - Products
    get:
      consumes:
      - application/json
      description: Mencari produk berdasarkan ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Ambil Detail Produk
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Mengubah data produk berdasarkan ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Data Update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Update Produk (Admin Only)
      tags:
      - Products
  /register:
    post:
      consumes:
      - application/json
      description: Input data user untuk disimpan ke database
      parameters:
      - description: Data User
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Mendaftarkan user baru
      tags:
      - Auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"